pyproject.toml
[tool.poetry]
name = "med-ifk-predict"
version = "0.1.0"
description = ""
authors = ["m.sultanov <m.sultanov@stcgeo.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^6.1.0"
black = "^23.11.0"
isort = "^5.12.0"
pytelegrambotapi = "^4.14.0"
pymongo = "^4.6.0"
aiohttp = "^3.9.1"
pyzmq = "^25.1.1"
telebot = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
